pyproject.toml
[tool.poetry]
name = "avocado-config-gen"
version = "0.1.0"
description = ""
authors = ["Edward George <edwardgeorge@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"ruamel.yaml" = "^0.17.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mockify = "^0.12.0"

[tool.poetry.scripts]
config-gen = "avocado_config_gen:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 118
target-version = ['py38']
